requirements.txt
# Core dependencies
openai>=1.0.0
asyncio-mqtt>=0.16.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0

# LLM and AI
tiktoken>=0.5.0
langchain>=0.1.0
anthropic>=0.7.0  # Optional: for Claude integration

# Data processing
pandas>=2.0.0
jsonschema>=4.0.0

# Async and concurrent processing
aiohttp>=3.8.0
aiofiles>=23.0.0

# Logging and monitoring
loguru>=0.7.0
rich>=13.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0